Unittest:dataLayer access to first item insted of last
In MongoDB it is possible to insert multiple items
with same metadata attributes, so to obtain the relevant
query item I use to call the last item.
since now the unittest run a tearup function that deletes all branches
that are related to the test in mongoDB we are only left with one
relevant query so their is no need to look up the last item.